Move estimatedTurns function to lib.
