+++ src/diet.ts
@@ -47,8 +47,8 @@ import {
 } from "libram";
 import { acquire } from "./acquire";
 import { embezzlerCount } from "./embezzlers";
-import { estimatedTurns, globalOptions } from "./globalvars";
-import { baseMeat, clamp, ensureEffect, setChoice } from "./lib";
+import { globalOptions } from "./globalvars";
+import { baseMeat, clamp, ensureEffect, estimatedTurns, setChoice } from "./lib";
 
 const MPA = get("valueOfAdventure");
 print(`Using adventure value ${MPA}.`, "blue");
