+++ src/globalvars.ts
@@ -1,7 +1,3 @@
-import { inebrietyLimit, myAdventures, myInebriety, myTurncount } from "kolmafia";
-import { $item, have } from "libram";
-import { embezzlerCount } from "./embezzlers";
-
 export const log = {
   initialEmbezzlersFought: 0,
   digitizedEmbezzlersFought: 0,
@@ -13,15 +9,3 @@ export const globalOptions: { ascending: boolean; stopTurncount: number | null; 
     ascending: false,
     noBarf: false,
   };
-
-export function estimatedTurns(): number {
-  let turns;
-  if (globalOptions.stopTurncount) turns = globalOptions.stopTurncount - myTurncount();
-  else if (globalOptions.noBarf) turns = embezzlerCount();
-  else
-    turns =
-      (myAdventures() + (globalOptions.ascending && myInebriety() <= inebrietyLimit() ? 60 : 0)) *
-      (have($item`mafia thumb ring`) ? 1.04 : 1);
-
-  return turns;
-}
